fix(listener): plug-n-play support
Prior to this commit, the joypad listener would only work if the joypad
was plugged in when the listener was started.

In this commit, modify the listener to still work if the joypad is plugged
in after the listener is started, and also if the joypad is unplugged and
plugged back in.